moving Unity Reference Project to "Unity3D/", correcting namespaces, adding external dependencies in Plugins folder, adding STL import
