+++ Unity3D/Assets/UrdfImporter/Editor/TransformExtensions.cs
@@ -1,0 +1,55 @@
+﻿/*
+© Siemens AG, 2017
+Author: Dr. Martin Bischoff (martin.bischoff@siemens.com)
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+<http://www.apache.org/licenses/LICENSE-2.0>.
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+using UnityEngine;
+
+namespace RosSharp.Urdf
+{
+    public static class TransformExtensions
+    {
+        public static void SetParentAndAlign(this Transform transform, Transform parent, bool keepLocalTransform = true)
+        {
+            Vector3 localPosition = transform.localPosition;
+            Quaternion localRotation = transform.localRotation;
+            transform.parent = parent;
+            if (keepLocalTransform)
+            {
+                transform.position = transform.parent.position + localPosition;
+                transform.rotation = transform.parent.rotation * localRotation;
+            }
+            else
+            {
+                transform.localPosition = Vector3.zero;
+                transform.localRotation = Quaternion.identity;
+            }
+        }
+
+        public static Transform FindDeepChild(this Transform parent, string name)
+        {
+            Transform result = parent.Find(name);
+            if (result != null)
+                return result;
+            foreach (Transform child in parent)
+            {
+                result = child.FindDeepChild(name);
+                if (result != null)
+                    return result;
+            }
+            return null;
+        }
+    }
+}
