--- UrdfImporter/UrdfSharpUnity3D/.gitignore
@@ -1,21 +1,0 @@
-# Select Unity3D UrdfImporter Assets only:
-
-# ## Ignore everything: ##
-
-* 
-
-# ## Except: ##
-
-!.gitignore
-!UrdfImporter.unitypackage
-
-!Assets/
-!Assets/UrdfImporter/
-!Assets/UrdfImporter/*/
-!Assets/UrdfImporter/*/*
-
-# ## But Ignore again: ##
-
-## Unity Meta file:
-
-*.meta
